requirements.txt
httplib2~=0.22.0
google-api-python-client~=2.166.0
flask~=3.1.0